chore: Fixes the erc20 revoke allowance tests
