+++ test/e2e/tests/confirmations/transactions/increase-token-allowance-redesign.spec.ts
@@ -2,8 +2,6 @@ import FixtureBuilder from '../../../fixture-builder';
 import {
   defaultGanacheOptions,
   defaultGanacheOptionsForType2Transactions,
-  largeDelayMs,
-  veryLargeDelayMs,
   WINDOW_TITLES,
   withFixtures,
 } from '../../../helpers';
@@ -12,7 +10,12 @@ import ContractAddressRegistry from '../../../seeder/contract-address-registry';
 import { SMART_CONTRACTS } from '../../../seeder/smart-contracts';
 import { Driver } from '../../../webdriver/driver';
 import { scrollAndConfirmAndAssertConfirm } from '../helpers';
-import { openDAppWithContract, TestSuiteArguments } from './shared';
+import {
+  assertChangedSpendingCap,
+  editSpendingCap,
+  openDAppWithContract,
+  TestSuiteArguments,
+} from './shared';
 
 describe('Confirmation Redesign ERC20 Increase Allowance', function () {
   describe('Submit an increase allowance transaction @no-mmi', function () {
@@ -157,43 +160,4 @@ export async function mocked4BytesIncreaseAllowance(mockServer: Mockttp) {
 async function createERC20IncreaseAllowanceTransaction(driver: Driver) {
   await driver.switchToWindowWithTitle(WINDOW_TITLES.TestDApp);
   await driver.clickElement('#increaseTokenAllowance');
-}
-
-export async function editSpendingCap(driver: Driver, newSpendingCap: string) {
-  await driver.switchToWindowWithTitle(WINDOW_TITLES.Dialog);
-  await driver.clickElement('[data-testid="edit-spending-cap-icon"');
-
-  await driver.fill(
-    '[data-testid="custom-spending-cap-input"]',
-    newSpendingCap,
-  );
-
-  await driver.delay(largeDelayMs);
-
-  await driver.clickElement({ text: 'Save', tag: 'button' });
-
-  // wait for the confirmation to be updated before submitting tx
-  await driver.delay(veryLargeDelayMs * 2);
-}
-
-export async function assertChangedSpendingCap(
-  driver: Driver,
-  newSpendingCap: string,
-) {
-  await driver.switchToWindowWithTitle(WINDOW_TITLES.ExtensionInFullScreenView);
-
-  await driver.clickElement({ text: 'Activity', tag: 'button' });
-
-  await driver.delay(veryLargeDelayMs);
-
-  await driver.clickElement(
-    '.transaction-list__completed-transactions .activity-list-item:nth-of-type(1)',
-  );
-
-  await driver.waitForSelector({
-    text: `${newSpendingCap} TST`,
-    tag: 'span',
-  });
-
-  await driver.waitForSelector({ text: 'Confirmed', tag: 'div' });
 }
