+++ test/e2e/tests/confirmations/transactions/revoke-allowance-redesign.spec.ts
@@ -3,12 +3,13 @@ import { MockttpServer } from 'mockttp';
 import { WINDOW_TITLES } from '../../../helpers';
 import { Driver } from '../../../webdriver/driver';
 import { scrollAndConfirmAndAssertConfirm } from '../helpers';
-import { mocked4BytesApprove } from './erc20-approve-redesign.spec';
 import {
   assertChangedSpendingCap,
   editSpendingCap,
-} from './increase-token-allowance-redesign.spec';
-import { openDAppWithContract, TestSuiteArguments } from './shared';
+  mocked4BytesApprove,
+  openDAppWithContract,
+  TestSuiteArguments,
+} from './shared';
 
 const {
   defaultGanacheOptions,
