+++ test/e2e/tests/confirmations/transactions/shared.ts
@@ -1,6 +1,6 @@
 /* eslint-disable @typescript-eslint/no-require-imports, @typescript-eslint/no-var-requires */
-import { MockedEndpoint } from 'mockttp';
-import { veryLargeDelayMs } from '../../../helpers';
+import { MockedEndpoint, MockttpServer } from 'mockttp';
+import { largeDelayMs, veryLargeDelayMs } from '../../../helpers';
 import { Ganache } from '../../../seeder/ganache';
 import ContractAddressRegistry from '../../../seeder/contract-address-registry';
 import { Driver } from '../../../webdriver/driver';
@@ -237,4 +237,67 @@ export async function assertAdvancedGasDetailsWithL2Breakdown(driver: Driver) {
   await driver.waitForSelector({ css: 'p', text: 'L2 fee' });
   await driver.waitForSelector({ css: 'p', text: 'Speed' });
   await driver.waitForSelector({ css: 'p', text: 'Max fee' });
+}
+
+export async function editSpendingCap(driver: Driver, newSpendingCap: string) {
+  await driver.switchToWindowWithTitle(WINDOW_TITLES.Dialog);
+  await driver.clickElement('[data-testid="edit-spending-cap-icon"');
+
+  await driver.fill(
+    '[data-testid="custom-spending-cap-input"]',
+    newSpendingCap,
+  );
+
+  await driver.delay(largeDelayMs);
+
+  await driver.clickElement({ text: 'Save', tag: 'button' });
+
+  // wait for the confirmation to be updated before submitting tx
+  await driver.delay(veryLargeDelayMs * 2);
+}
+
+export async function assertChangedSpendingCap(
+  driver: Driver,
+  newSpendingCap: string,
+) {
+  await driver.switchToWindowWithTitle(WINDOW_TITLES.ExtensionInFullScreenView);
+
+  await driver.clickElement({ text: 'Activity', tag: 'button' });
+
+  await driver.delay(veryLargeDelayMs);
+
+  await driver.clickElement(
+    '.transaction-list__completed-transactions .activity-list-item:nth-of-type(1)',
+  );
+
+  await driver.waitForSelector({
+    text: `${newSpendingCap} TST`,
+    tag: 'span',
+  });
+
+  await driver.waitForSelector({ text: 'Confirmed', tag: 'div' });
+}
+
+export async function mocked4BytesApprove(mockServer: MockttpServer) {
+  return await mockServer
+    .forGet('https://www.4byte.directory/api/v1/signatures/')
+    .always()
+    .withQuery({ hex_signature: '0x095ea7b3' })
+    .thenCallback(() => ({
+      statusCode: 200,
+      json: {
+        count: 1,
+        next: null,
+        previous: null,
+        results: [
+          {
+            id: 149,
+            created_at: '2016-07-09T03:58:29.617584Z',
+            text_signature: 'approve(address,uint256)',
+            hex_signature: '0x095ea7b3',
+            bytes_signature: '\t^§³',
+          },
+        ],
+      },
+    }));
 }
